requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
pytest>=6.2.4,<6.3
pytest-mock>=3.6.1,<3.7
pytest-django>=4.4.0,<4.5
pytest-cov>=2.12.1,<2.13
drf-spectacular>=0.15.1,<0.16
mixer>=7.2.2,<7.3